move device tests for linux to separate file
